refactor: relocate production-related files into module directory
