+++ app/production/models.py
@@ -1,4 +1,5 @@
 from typing import Optional
+
 from sqlmodel import Field, SQLModel
